feat(server): sqlx (#413)
* feat(server): add the `db` module

* feat(server): use sqlx in the media module

* feat(server): use sqlx in the users module

* feat(server): use sqlx in the info module

* feat(server): sqlx prepare

* fix(ci): style check

* refactor(server): clippy

* feat(server): use sqlx

* style(server): pgFormatter

* chore(server): don't connect databases in tests

* fix(server): move generated files from sqlx

* test(server): move cache test

* style(server): format SQLs of the users module

* refactor: remove the testing database

* chore(server): sqlx prepare

* feat(server): database error handling

* refactor(server): reuse the database connection

* fix(server): adjust the order of the fields to match the database order

* feat(server): database check

* style(server): rename unused variables

* fix(server): database check

* refactor(server): clippy

* feat(server): check schema

* feat(server): do not check the storage API under the CI

* fix(server): order in the schema

* fix(server): adjust the order of the spaces columns

+++ apps/server/src/messages/api.rs
@@ -63,5 +63,5 @@ pub struct MoveMessageBetween {
 pub struct GetMessagesByChannel {
     pub channel_id: Uuid,
     pub before: Option<f64>,
-    pub limit: Option<i32>,
+    pub limit: Option<i64>,
 }
